pyproject.toml
[tool.poetry]
name = "atd_00_template_api_python"
version = "0.0.1"
description = ""
readme = "README.md"
authors = ["mountain-gorilla-inc"]

[tool.poetry.dependencies]
python = "~3.9"
fastapi = "^0.76.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
gunicorn = "^20.1.0"
invoke = "^1.7.0"
mysqlclient = "^2.1.0"
SQLAlchemy = "^1.4.36"
alembic = "^1.7.7"
pydantic = {extras = ["email"], version = "^1.9.0"}
python-dotenv = "^0.20.0"
PyYAML = "^6.0"
cognito-pyauth = "^1.3.2"

[tool.poetry.dev-dependencies]
mypy = "^0.950"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
requests = "^2.27.1"

[tool.mypy]
python_version = 3.9
plugins = "pydantic.mypy"
check_untyped_defs = true
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
